+++ R/Microsoft365R.R
@@ -20,7 +20,18 @@ utils::globalVariables(c("self", "private"))
     register_graph_class("list", ms_list,
         function(props) !is_empty(props$list))
 
-    add_methods()
+    register_graph_class("team", ms_team,
+        function(props) "memberSettings" %in% names(props))
+
+    register_graph_class("channel", ms_channel,
+        function(props) "moderationSettings" %in% names(props))
+
+    register_graph_class("chatMessage", ms_chat_message,
+        function(props) "body" %in% names(props) && "messageType" %in% names(props))
+
+    add_graph_methods()
+    add_user_methods()
+    add_group_methods()
 }
 
 # default app ID